+++ backend/src/main/java/com/sabujaks/irs/domain/system/model/entity/BaseInfo.java
@@ -1,4 +1,4 @@
-package com.sabujaks.irs.domain.data_init.model.entity;
+package com.sabujaks.irs.domain.system.model.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
