+++ backend/src/main/java/com/sabujaks/irs/domain/system/model/response/CategoryRes.java
@@ -1,4 +1,4 @@
-package com.sabujaks.irs.domain.data_init.model.response;
+package com.sabujaks.irs.domain.system.model.response;
 
 import lombok.*;
