+++ cicd/k8s/backend/backend-deployment-blue.yml
@@ -25,14 +25,14 @@ spec:
                 name: backend-cm
           readinessProbe:
             httpGet:
-              path: /api/total-process/ex01
+              path: /api/system-health
               port: 8080
             initialDelaySeconds: 20
             periodSeconds: 10
             successThreshold: 3
           livenessProbe:
             httpGet:
-              path: /api/total-process/ex01
+              path: /api/system-health
               port: 8080
             initialDelaySeconds: 20
             periodSeconds: 10
